Refactor and improve code consistency across modules
Replaced manual array copying with System.arraycopy in ChessGame, simplified switch statement, and improved test assertions in FindLettersInOrderTest. Refactored method signatures and usages for clarity in CoinThrowing, Move, MoveGenerator, and WordsWithSameLetters. Added missing null checks and assertions, improved serialization annotations, and fixed type casting in ImageEncoder. Enhanced equals/hashCode/toString in Layer, removed unnecessary rounding in Vector2I, and cleaned up gray level calculation in SquareCircleClassification. Removed unnecessary exception from minMax in TicTacToe.

+++ src/ImageEncryption/ImageEncoder.java
@@ -66,7 +66,7 @@ public class ImageEncoder {
                         }
                     }
                     curr = curr & 3;
-                    bytes[i] += curr << j;
+                    bytes[i] += (byte) (curr << j);
                 }
             }
         }
